separated dependency injector
