+++ app/src/main/java/es/npatarino/android/gotchallenge/common/di/components/ActivityComponent.java
@@ -1,0 +1,13 @@
+package es.npatarino.android.gotchallenge.common.di.components;
+
+import android.content.Context;
+
+import dagger.Component;
+import es.npatarino.android.gotchallenge.common.di.Activity;
+import es.npatarino.android.gotchallenge.common.di.modules.ActivityModule;
+
+@Activity
+@Component(dependencies = AppComponent.class, modules = ActivityModule.class)
+public interface ActivityComponent {
+    Context context();
+}
