+++ app/src/main/java/es/npatarino/android/gotchallenge/common/di/modules/ActivityModule.java
@@ -1,10 +1,10 @@
-package es.npatarino.android.gotchallenge.di.modules;
+package es.npatarino.android.gotchallenge.common.di.modules;
 
 import android.content.Context;
 
 import dagger.Module;
 import dagger.Provides;
-import es.npatarino.android.gotchallenge.di.Activity;
+import es.npatarino.android.gotchallenge.common.di.Activity;
 
 @Module public class ActivityModule {
     private final Context mContext;
