--- app/src/main/java/es/npatarino/android/gotchallenge/di/components/ActivityComponent.java
@@ -1,14 +1,0 @@
-package es.npatarino.android.gotchallenge.di.components;
-
-import android.content.Context;
-
-import dagger.Component;
-import es.npatarino.android.gotchallenge.di.Activity;
-import es.npatarino.android.gotchallenge.di.AppComponent;
-import es.npatarino.android.gotchallenge.di.modules.ActivityModule;
-
-@Activity
-@Component(dependencies = AppComponent.class, modules = ActivityModule.class)
-public interface ActivityComponent {
-    Context context();
-}
